se ve mas boniko, hice separacion de funciones para claridad en codigo
