+++ static/css/index.css
@@ -45,7 +45,7 @@ body{
     margin: 20px;
     width: 95%;
     height: 40rem;
-    background-color: rgb(204, 48, 48);
+    background-color: rgb(204, 54, 54);
     border-radius: 0px 0px 20px 20px;
     box-shadow: 10px 8px rgb(122, 29, 29);
 }
@@ -68,10 +68,12 @@ body{
 }
 .poke_list_element p{
     position: relative;
+    transition: all ease 500ms;
 }
 .sprite{
     background-color: rgb(245, 244, 244);
     width: 8rem;
+    height: 8rem;
     border-radius: 0px 0px 5px 5px;
     border: 5px inset ;
 }
@@ -79,7 +81,7 @@ body{
     margin-left: 10px;
 }
 .zone_select{
-    background-color: yellow;
+    background-color: rgb(248, 203, 192);
     border-radius: 0px 0px 0px 20px;
 
     width: 100px;
@@ -87,6 +89,20 @@ body{
     position: absolute;
     bottom: 0;
 }
+
+
+.zone_select .poke_list_element{
+    padding: 0;
+    text-align: center;
+    margin: 5px 0 0 2px;
+    width: 80px;
+    height: 50px;
+    position: absolute;
+    bottom: 10px;
+    font-size: 13px;
+    /* animation: ; */
+}
+
 .dex_data{
     color: rgb(36, 36, 36);
     background-color: rgb(51, 51, 51);
@@ -96,4 +112,14 @@ body{
     /* box-shadow: 5px 5px 5px rgb(37, 37, 37,.5); */
     border: 5px inset ;
     color: white;
+}
+
+.info_zone{
+    background-color: rgb(167, 39, 39);
+    box-shadow: 10px 2px rgb(197, 64, 64);
+    width: fit-content;
+    padding-right: 2rem;
+    height: 100%;
+    border-radius: 0px 3px 3px 20px;
+
 }
